aicircle-main/aicircle_back/organizations/serializers.py
from rest_framework import serializers

from . import models


class OrganizationSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Organization
        fields = "__all__"
        read_only_fields = ["id", "created_at", "updated_at"]
