aicircle-main/aicircle_back/pyproject.toml
[project]
name = "aicircle_back"
version = "0.1.0"
description = ""
authors = [
    {name = "dishantsethi",email = "dishantsethi14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "django (==5.2.1)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-environ (==0.9.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "boto3 (==1.34.153)",
    "celery (==5.2.7)",
    "channels (==3.0.5)",
    "django-redis (==5.2.0)",
    "django-ses (==4.4.0)",
    "django-versatileimagefield (==2.2)",
    "ipdb (==0.13.13)",
    "kombu (==5.2.4)",
    "django-celery-beat (==2.8.0)",
    "django-log-request-id (>=2.1.0,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "django-sites (>=0.11,<0.12)",
    "django-mail-templated (>=2.6.5,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "django-timezone-field (>=7.1,<8.0)",
    "flower (>=2.0.1,<3.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "django-debug-toolbar (>=6.1.0,<7.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "livekit (>=1.0.20,<2.0.0)",
    "livekit-api (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
